Changed name in static_list to be an attribute instead of an initialization parameter. Updated code, docs, and tests

+++ lib/urbanairship/devices/static_lists.rb
@@ -7,15 +7,15 @@ module Urbanairship
     class StaticList
       include Urbanairship::Common
       include Urbanairship::Loggable
-
-      def initialize(client: required('client'), name: required('name'))
-        fail ArgumentError,
-           'Neither client nor name can be set to nil' if client.nil? or name.nil?
+      attr_accessor :name
+      def initialize(client: required('client'))
+        fail ArgumentError, 'Client cannot be set to nil' if client.nil?
         @client = client
-        @name = name
+        @name = nil
       end
 
       def create(description: nil, extras: nil)
+        fail ArgumentError, 'Name must be set' if @name.nil?
         payload = {'name' => @name}
         payload['description'] = description unless description.nil?
         payload['extras'] = extras unless extras.nil?
@@ -31,6 +31,7 @@ module Urbanairship
       end
 
       def upload(csv_file: required('csv_file'), gzip: false)
+        fail ArgumentError, 'Name must be set' if @name.nil?
         if gzip
           response = @client.send_request(
               method: 'PUT',
@@ -52,6 +53,7 @@ module Urbanairship
       end
 
       def update(description: nil, extras: nil)
+        fail ArgumentError, 'Name must be set' if @name.nil?
         fail ArgumentError,
            'Either description or extras must be set to a value' if description.nil? and extras.nil?
         payload = {}
@@ -68,6 +70,7 @@ module Urbanairship
       end
 
       def lookup
+        fail ArgumentError, 'Name must be set' if @name.nil?
         response = @client.send_request(
           method: 'GET',
           url: LISTS_URL + @name
@@ -77,6 +80,7 @@ module Urbanairship
       end
 
       def delete
+        fail ArgumentError, 'Name must be set' if @name.nil?
         response = @client.send_request(
           method: 'DELETE',
           url: LISTS_URL + @name
